samples/requirements_interop_cu12.txt
# SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
# SPDX-License-Identifier: Apache-2.0
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
# http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

# CUDA 12-specific Python packages for CV-CUDA interoperability samples

# Include common interop requirements
-r requirements_interop_common.txt

# PyTorch for CUDA 12.1 (only torch, interop samples don't need torchvision)
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.8.0

# CUDA 12-specific packages
--extra-index-url https://pypi.nvidia.com/
cuda-python==12.9.4
nvidia-nvimgcodec-cu12==0.6.1.37
cupy-cuda12x==13.6.0

# CV-CUDA for CUDA 12
cvcuda-cu12
